SGD-SVM with polynomial feature extraction using gpu
